[8] Major refactor to clean up generation code
- Created a package for docker compose file generation and split up
GenerateTask
- Merge together redundant functions for readability
- Grouped indented functions under nested 'for' statements so you can
tell by the code whether something is indented or not
