+++ src/test/scala/au/com/agiledigital/toolform/app/GenerateTaskTest.scala
@@ -22,50 +22,53 @@ class GenerateTaskTest extends FlatSpec with Matchers {
     outputFile.deleteOnExit()
     ToolFormApp.execute(List("generate", "-i", testFile.getAbsolutePath, "-o", outputFile.getAbsolutePath).toArray)
     val result = Source.fromFile(outputFile.getAbsolutePath).mkString
+    print("-------")
+    print(result)
+    print("-------")
     result should equal(s"""# Generated by ${BuildInfo.name} (${BuildInfo.version})
           |version: '3'
-          |  services:
-          |    seswipmainpublicapinginx:
-          |      image: se_swip_main_public_api_nginx
-          |      restart: always
-          |      ports:
-          |      - "80:80"
-          |    clientpublic:
-          |      image: se_swip/client/public
-          |      restart: always
-          |      labels:
-          |        source.path: "client/public"
-          |        project.artefact: "true"
-          |    publicapi:
-          |      image: se_swip/public_api
-          |      restart: always
-          |      labels:
-          |        source.path: "server/public"
-          |        project.artefact: "true"
-          |    seswipelasticsearch:
-          |      image: se_swip/se_swip_elastic_search
-          |      restart: always
-          |      labels:
-          |        source.path: "resources/elastic_search"
-          |        project.artefact: "true"
-          |    seswipinfluxdb:
-          |      image: se_swip/se_swip_influx_db
-          |      restart: always
-          |      labels:
-          |        source.path: "resources/influxdb"
-          |        project.artefact: "true"
-          |    seswipcarbon:
-          |      image: dockerana/carbon
-          |      restart: always
-          |    seswipdb:
-          |      image: docker.agiledigital.com.au:5000/agile/pgpool2
-          |      restart: always
-          |      environment:
-          |      - BACKEND_PORT=5432
-          |      - BACKEND_HOST=localhost
-          |    seswipmailrelay:
-          |      image: djfarrelly/maildev
-          |      restart: always
+          |services:
+          |  seswipmainpublicapinginx:
+          |    image: se_swip_main_public_api_nginx
+          |    restart: always
+          |    ports:
+          |    - "80:80"
+          |  clientpublic:
+          |    image: se_swip/client/public
+          |    restart: always
+          |    labels:
+          |      source.path: "client/public"
+          |      project.artefact: "true"
+          |  publicapi:
+          |    image: se_swip/public_api
+          |    restart: always
+          |    labels:
+          |      source.path: "server/public"
+          |      project.artefact: "true"
+          |  seswipelasticsearch:
+          |    image: se_swip/se_swip_elastic_search
+          |    restart: always
+          |    labels:
+          |      source.path: "resources/elastic_search"
+          |      project.artefact: "true"
+          |  seswipinfluxdb:
+          |    image: se_swip/se_swip_influx_db
+          |    restart: always
+          |    labels:
+          |      source.path: "resources/influxdb"
+          |      project.artefact: "true"
+          |  seswipcarbon:
+          |    image: dockerana/carbon
+          |    restart: always
+          |  seswipdb:
+          |    image: docker.agiledigital.com.au:5000/agile/pgpool2
+          |    restart: always
+          |    environment:
+          |    - BACKEND_PORT=5432
+          |    - BACKEND_HOST=localhost
+          |  seswipmailrelay:
+          |    image: djfarrelly/maildev
+          |    restart: always
           |""".stripMargin)
   }
 }
